Minor cleanups in account manager
